refactor: Move AppUnavailableScreen in the ui package
Resolves: DCMAW-9827
