+++ app/src/androidTest/java/uk/gov/onelogin/ui/error/unavailable/AppUnavailableScreenTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.feature.unavailable.ui
+package uk.gov.onelogin.ui.error.unavailable
 
 import androidx.test.espresso.Espresso
 import androidx.test.espresso.NoActivityResumedException
