+++ app/src/androidTest/java/uk/gov/onelogin/ui/error/unavailable/UnavailableAnalyticsViewModelTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.feature.unavailable.ui
+package uk.gov.onelogin.ui.error.unavailable
 
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
