+++ app/src/main/java/uk/gov/onelogin/ui/error/ErrorGraphObject.kt
@@ -8,7 +8,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
 import androidx.navigation.navigation
-import uk.gov.onelogin.feature.unavailable.ui.AppUnavailableScreen
+import uk.gov.onelogin.ui.error.unavailable.AppUnavailableScreen
 import uk.gov.onelogin.ui.error.update.UpdateRequiredScreen
 
 object ErrorGraphObject {
